requirements.txt
pyyaml==6.0
pandas==1.5.0
scikit-learn==1.1.2
hyperopt==0.2.7